requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pandas==2.1.3
networkx==3.2
rapidfuzz==3.5.2
sentence-transformers==2.2.2
pytesseract==0.3.10
pdfplumber==0.10.3
fpdf==1.7.2
Pillow==10.1.0
python-multipart==0.0.6
requests==2.31.0
